Add a new IPv4.All table that enumerates IPv4 packets received by device, source and destination address and the IPv4 protocol value. This table can be considered a parent table to the existing (unchanged) IPv4.UDP table which adds the detail of destination port for IPv4 packets containing a UDP datagram.
Move context initialization logic to the individual readDevice
threads. The change allows handling of network devices that are not
present at start up but appear later which happens in some race
cases. The logic also allows the probe to report statistics for the
capture devices that are present even in the case that one is not.